rename-collection-repository: rename list to collection repository
