+++ core/src/main/java/org/hisp/dhis/android/core/relationship/RelationshipModule.java
@@ -27,7 +27,7 @@
  */
 package org.hisp.dhis.android.core.relationship;
 
-import org.hisp.dhis.android.core.arch.repositories.ReadOnlyListRepository;
+import org.hisp.dhis.android.core.arch.repositories.ReadOnlyCollectionRepository;
 import org.hisp.dhis.android.core.data.database.DatabaseAdapter;
 
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
@@ -35,11 +35,11 @@ import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 @SuppressFBWarnings("URF_UNREAD_PUBLIC_OR_PROTECTED_FIELD")
 public final class RelationshipModule {
 
-    public final ReadOnlyListRepository<RelationshipType> relationshipType;
+    public final ReadOnlyCollectionRepository<RelationshipType> relationshipType;
 
     public final RelationshipRepositoryInterface relationship;
 
-    private RelationshipModule(ReadOnlyListRepository<RelationshipType> relationshipTypeRepository,
+    private RelationshipModule(ReadOnlyCollectionRepository<RelationshipType> relationshipTypeRepository,
                                RelationshipRepositoryInterface relationshipRepository) {
         this.relationshipType = relationshipTypeRepository;
         this.relationship = relationshipRepository;
