+++ core/src/main/java/org/hisp/dhis/android/core/relationship/RelationshipTypeRepository.java
@@ -27,20 +27,20 @@
  */
 package org.hisp.dhis.android.core.relationship;
 
-import org.hisp.dhis.android.core.arch.repositories.ReadOnlyListRepository;
-import org.hisp.dhis.android.core.arch.repositories.ReadOnlyListRepositoryImpl;
+import org.hisp.dhis.android.core.arch.repositories.ReadOnlyCollectionRepository;
+import org.hisp.dhis.android.core.arch.repositories.ReadOnlyCollectionRepositoryImpl;
 import org.hisp.dhis.android.core.data.database.DatabaseAdapter;
 
 import java.util.HashSet;
 import java.util.Set;
 
-public final class RelationshipTypeRepository implements ReadOnlyListRepository<RelationshipType> {
+public final class RelationshipTypeRepository implements ReadOnlyCollectionRepository<RelationshipType> {
 
-    private final ReadOnlyListRepository<RelationshipConstraint> constraintRepository;
-    private final ReadOnlyListRepository<RelationshipType> rawTypeRepository;
+    private final ReadOnlyCollectionRepository<RelationshipConstraint> constraintRepository;
+    private final ReadOnlyCollectionRepository<RelationshipType> rawTypeRepository;
 
-    private RelationshipTypeRepository(ReadOnlyListRepository<RelationshipConstraint> constraintRepository,
-                               ReadOnlyListRepository<RelationshipType> rawTypeRepository) {
+    private RelationshipTypeRepository(ReadOnlyCollectionRepository<RelationshipConstraint> constraintRepository,
+                               ReadOnlyCollectionRepository<RelationshipType> rawTypeRepository) {
         this.constraintRepository = constraintRepository;
         this.rawTypeRepository = rawTypeRepository;
     }
@@ -61,11 +61,11 @@ public final class RelationshipTypeRepository implements ReadOnlyListRepository<
 
     static RelationshipTypeRepository create(DatabaseAdapter databaseAdapter) {
         return new RelationshipTypeRepository(
-                new ReadOnlyListRepositoryImpl<>(
+                new ReadOnlyCollectionRepositoryImpl<>(
                         RelationshipConstraintStore.create(databaseAdapter),
                         RelationshipConstraint.factory
                 ),
-                new ReadOnlyListRepositoryImpl<>(
+                new ReadOnlyCollectionRepositoryImpl<>(
                         RelationshipTypeStore.create(databaseAdapter),
                         RelationshipType.factory
                 )
